Bump to 3.3.0, Update updater, popup NavKit settings when paths are not set

+++ src/module/PersistedSettings.cpp
@@ -25,6 +25,8 @@ void PersistedSettings::load() {
         Logger::log(NK_INFO, "Old settings file found. Merging settings...");
         CSimpleIniA oldIni;
         oldIni.SetUnicode();
+        bool mergedSettingsSaved = false;
+
         if (oldIni.LoadFile(oldIniPath.c_str()) >= 0) {
             if (ini.LoadFile(iniPath.c_str()) < 0) {
                 Logger::log(NK_ERROR, "New NavKit.ini not found. Creating a new one from old settings.");
@@ -33,20 +35,6 @@ void PersistedSettings::load() {
             CSimpleIniA::TNamesDepend sections;
             oldIni.GetAllSections(sections);
 
-            if (const auto backgroundColor = oldIni.GetValue("NavKit", "backgroundColor", ""); backgroundColor[0] !=
-                '\0') {
-                ini.SetValue("NavKit", "backgroundColor", backgroundColor);
-            }
-            if (const auto hitman = oldIni.GetValue("NavKit", "hitman", ""); hitman[0] != '\0') {
-                ini.SetValue("NavKit", "hitman", hitman);
-            }
-            if (const auto output = oldIni.GetValue("NavKit", "output", ""); output[0] != '\0') {
-                ini.SetValue("NavKit", "output", output);
-            }
-            if (const auto blender = oldIni.GetValue("NavKit", "blender", ""); blender[0] != '\0') {
-                ini.SetValue("NavKit", "blender", blender);
-            }
-            // Merge from old settings that match the new settings schema
             for (const auto& section : sections) {
                 if (const CSimpleIniA::TKeyVal* keys = oldIni.GetSection(section.pItem)) {
                     for (auto keyval = keys->begin(); keyval != keys->end(); ++keyval) {
@@ -54,6 +42,13 @@ void PersistedSettings::load() {
                         if (const auto oldValue = oldIni.GetValue(section.pItem, key, "")) {
                             if (oldValue[0] != '\0') {
                                 ini.SetValue(section.pItem, key, oldValue);
+                                Logger::log(
+                                    NK_INFO,
+                                    "Transferred setting [%s] %s = %s",
+                                    section.pItem,
+                                    key,
+                                    oldValue
+                                );
                             }
                         }
                     }
@@ -61,18 +56,26 @@ void PersistedSettings::load() {
             }
 
             Logger::log(NK_INFO, "Settings merged. Saving updated NavKit.ini.");
-            save();
+            mergedSettingsSaved = save();
         } else {
             Logger::log(
                 NK_ERROR, "Failed to load NavKit.ini.old for merging. Loading current settings if they exist.");
             ini.LoadFile(iniPath.c_str());
         }
 
-        try {
-            std::filesystem::remove(oldIniPath);
-            Logger::log(NK_INFO, "Removed temporary old settings file.");
-        } catch (const std::filesystem::filesystem_error& e) {
-            Logger::log(NK_ERROR, "Failed to delete NavKit.ini.old: %s", e.what());
+        if (mergedSettingsSaved) {
+            try {
+                std::filesystem::remove(oldIniPath);
+                Logger::log(NK_INFO, "Removed temporary old settings file.");
+            } catch (const std::filesystem::filesystem_error& e) {
+                Logger::log(NK_ERROR, "Failed to delete NavKit.ini.old: %s", e.what());
+            }
+        } else {
+            Logger::log(
+                NK_ERROR,
+                "Merged settings were not saved successfully. Keeping NavKit.ini.old at %s.",
+                oldIniPath.c_str()
+            );
         }
     } else {
         ini.LoadFile(iniPath.c_str());
@@ -90,10 +93,13 @@ void PersistedSettings::load() {
     RecastAdapter::getInstance().loadSettings();
 }
 
-void PersistedSettings::save() const {
+bool PersistedSettings::save() const {
     if (ini.SaveFile(iniPath.c_str()) == SI_FAIL) {
         Logger::log(NK_ERROR, "Error saving settings file to %s", iniPath.c_str());
+        return false;
     }
+
+    return true;
 }
 
 const char* PersistedSettings::getValue(const std::string& folder, const std::string& key,
